requirements.txt
cbor2==5.6.5
deprecation==2.1.0
eyeD3==0.9.8
filetype==1.2.0
luma.core==2.4.2
luma.oled==3.14.0
mutagen==1.47.0
packaging==25.0
pillow==11.3.0
pyftdi==0.56.0
pygame==2.6.1
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
smbus2==0.5.0
spidev==3.7
